Add files via upload
launch_button.launch helps with ease in running the GUI. service_file controls what the button is going to return to the terminal, button_tester.py is the GUI python code utilizing PYQT5. After running launch, there should be a display with a red and blue button, and when either is pressed the terminal should return "1".